Move over collection tests to a new Document+CollectionTests.swift
